fix: bypass_device_restrictions test to test_testing.py
